requirements.txt
Django==5.0.2
python-dotenv==1.0.0
azure-storage-blob==12.19.0
Pillow==10.1.0 